# docs/handoff/01_postanovka/04_openapi.yaml
openapi: 3.1.0
info:
  title: reHome Knowledge Base API
  version: 1.0.0
  summary: API модуля базы знаний платформы reHome
  description: "Публичный API модуля базы знаний reHome, через который платформа rehome.one\nи другие потребители получают\
    \ доступ к статьям help-центра, документам,\nAI-чату, поиску и карточкам квартир.\n\nПолное описание архитектуры и принципов\
    \ — в документе\n«API базы знаний reHome — ТЗ v1.0».\n\n## Аутентификация\nПоддерживаются два режима:\n- **m2m** (server-to-server):\
    \ OAuth 2.0 Client Credentials через kb-auth Keycloak,\n  Authorization: Bearer <jwt>\n- **user** (browser-to-server):\
    \ HttpOnly Cookie kb_session с JWT,\n  устанавливается при логине пользователя в rehome.one\n\n## Контроль доступа\nКаждый\
    \ ресурс имеет access_level (PUBLIC / LOGGED / AGENT / STAFF / LEGAL / HR_RESTRICTED).\nscope пользователя вычисляется\
    \ бэкендом из проверенного токена,\nНИКОГДА не передаётся клиентом. Фильтрация применяется на уровне хранилища.\n\n##\
    \ Версионирование\nВерсия в URL: /api/v1/. Breaking changes требуют новой мажорной версии\nи параллельной работы старой\
    \ минимум 6 месяцев.\n"
  contact:
    name: reHome Engineering
    email: engineering@rehome.one
    url: https://rehome.one
  license:
    name: Proprietary
  termsOfService: https://rehome.one/legal/api-terms
servers:
- url: https://api.rehome.one/kb
  description: Production
- url: https://api.staging.rehome.one/kb
  description: Staging
...
security:
- BearerAuth: []
- CookieAuth: []
...
paths:
...
  /api/v1/admin/system-config:
    get:
      tags:
      - Admin
      summary: Системные настройки kb-модуля
      description: 'Конфигурация на лету: rate limits, LLM-провайдер, фичефлаги, лимиты модерации.

        '
      operationId: getSystemConfig
      x-implementation-status: implemented
      x-implementation-pr: 229
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/SystemConfig'
    patch:
      tags:
      - Admin
      summary: Изменение настроек
      operationId: updateSystemConfig
      x-implementation-status: implemented
      x-implementation-pr: 264
      x-implementation-note: ADR-0019; X-MFA-Token honest stub
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Patch-объект с полями SystemConfig
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/SystemConfig'
...
  /api/v1/admin/llm/active:
    put:
      tags:
      - Admin
      summary: Переключить активного LLM-провайдера
      description: 'Меняет провайдера без деплоя. Запрос требует подтверждение через MFA-челлендж.

        '
      operationId: setActiveLlmProvider
      x-implementation-status: implemented
      x-implementation-pr: 264
      x-implementation-note: ADR-0019 short-hand для PATCH system-config llm_provider key
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      - name: X-MFA-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - provider_id
              properties:
                provider_id:
                  type: string
                reason:
                  type: string
      responses:
        '200':
          description: Переключено
...
  /api/v1/admin/stats:
    get:
      tags:
      - Admin
      summary: Сводная статистика kb-модуля
      description: 'Метрики по запросам, чату, документам, инцидентам.

        Окно — последние 30 дней по умолчанию.

        '
      operationId: getAdminStats
      x-implementation-status: implemented
      x-implementation-pr: 227
      parameters:
      - name: from
        in: query
        schema:
          type: string
          format: date
      - name: to
        in: query
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ResponseEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/AdminStats'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Bearer JWT — m2m или user токен.

        m2m: получить через kb-auth Client Credentials Grant.

        '
    CookieAuth:
      type: apiKey
      in: cookie
      name: kb_session
      description: HttpOnly Secure cookie, ставится при логине в rehome.one
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      schema:
        type: string
      description: CSRF-токен (double-submit). Обязателен для browser-mutations.
...
  schemas:
...
    Meta:
      type: object
      required:
      - request_id
      - api_version
      - timestamp
      properties:
        request_id:
          type: string
          format: uuid
        api_version:
          type: string
          example: 1.0.0
        timestamp:
          type: string
          format: date-time
...
    ResponseEnvelope:
      type: object
      required:
      - meta
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
...
    SystemConfig:
      type: object
      properties:
        rate_limits:
          type: object
          properties:
            guest_per_minute:
              type: integer
            user_per_minute:
              type: integer
            m2m_per_minute:
              type: integer
            chat_messages_per_5min:
              type: integer
        feature_flags:
          type: object
          additionalProperties:
            type: boolean
          description: 'Например: chat_enabled, internal_chat_enabled, partner_webhooks_enabled'
        llm_config:
          type: object
          properties:
            active_provider:
              type: string
            fallback_provider:
              type: string
            ab_test_split_percent:
              type: integer
              minimum: 0
              maximum: 100
            max_context_tokens:
              type: integer
            temperature:
              type: number
        moderation:
          type: object
          properties:
            auto_publish_threshold:
              type: number
            require_review_for_categories:
              type: array
              items:
                type: string
        webhooks:
          type: object
          properties:
            max_retries:
              type: integer
            timeout_seconds:
              type: integer
...
    AdminStats:
      type: object
      properties:
        period:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
        requests:
          type: object
          properties:
            total:
              type: integer
            by_endpoint:
              type: object
              additionalProperties:
                type: integer
            by_status:
              type: object
              additionalProperties:
                type: integer
            error_rate_percent:
              type: number
        chat:
          type: object
          properties:
            sessions:
              type: integer
            messages:
              type: integer
            containment_rate:
              type: number
              description: Доля сессий без эскалации
            avg_rating:
              type: number
            no_answer_count:
              type: integer
            escalations:
              type: integer
        content:
          type: object
          properties:
            total_articles:
              type: integer
            total_documents:
              type: integer
            pending_reviews:
              type: integer
        security:
          type: object
          properties:
            open_incidents:
              type: integer
            critical_incidents:
              type: integer
            overdue_pd_requests:
              type: integer
...